Updated setup script

+++ setup.sh
@@ -1,43 +1,57 @@
 #!/bin/bash
 #Setup script for Xavier
 
-if [ -z %1]; then
+DIR="$(pwd)"
+if [[ $DIR == *"/go/src/"* ]]; then
+	echo "Error! Cannot execute setup script from this location! Exiting."
+	exit
+fi
+
+if [ -z %1 ]; then
 	echo "Error! No GitHub username was provided"
 	exit
 else
-	$USER = $1
+	GITUSER=$1
 fi 
 
 if [ -z "$GOPATH"]; then
-	export $GOPATH = $HOME/go
+	export $GOPATH=$HOME/go
 fi
 
-DIR = "$(pwd)"
-
-if [ ! -d "$GOPATH"]; then
+if [ ! -d "$GOPATH" ]; then
 	mkdir "$GOPATH/src/github.com/$USER/Xavier" && cd "$_"
 	git clone https://github.com/awarrier99/Xavier ./
 	rm -rf "$DIR"
 else
 	cd "$GOPATH"
-	if [ ! -d "/src"]; then
+	if [ ! -d "/src" ]; then
 		mkdir src
 	fi
 	cd src
 
-	if [ ! -d "/github.com"]; then
+	if [ ! -d "/github.com" ]; then
 		mkdir github.com
 	fi
 	cd github.com
 
-	if [ ! -d "/$USER"]; then
-		mkdir $USER
+	if [ ! -d "/$GITUSER" ]; then
+		mkdir $GITUSER
 	fi
-	cd $USER
+	cd $GITUSER
 
-	if [ -d "Xavier"]; then
-		echo "Error! Directory Xavier already exists!"
-		exit
+	if [ -d "Xavier" ]; then
+		echo "Directory Xavier already exists!"
+		echo "Do you want to overwrite its contents?"
+		read res
+		if [ $res == 'y' ] || [ $res == 'Y' ]; then
+			rm -rf "Xavier"
+			mkdir "Xavier"
+		else
+			echo "Exiting setup script."
+			exit
+		fi
+	else
+		mkdir "Xavier"
 	fi
 	cd Xavier
